chore: move charts to 3 main folders
