+++ src/pages/dashboard/AllLineschart/AllLinesChart.tsx
@@ -2,7 +2,7 @@ import { Tooltip } from '@mui/material'
 import { Skeleton } from 'antd'
 import { Fragment } from 'react'
 import { TEXTS } from 'src/resources/texts'
-import OperatorHbarChart from '../OperatorHbarChart/OperatorHbarChart'
+import OperatorHbarChart from './OperatorHbarChart/OperatorHbarChart'
 import { GroupByRes, useGroupBy } from 'src/api/groupByService'
 import { FC } from 'react'
 import { Moment } from 'moment/moment'
