+++ src/pages/dashboard/AllLineschart/OperatorHbarChart/operatorsHbarChart.scss
