+++ src/pages/dashboard/WorstLinesChart/LineHbarChart/HbarChart/HbarChart.scss
